Added textdata for adding text
