+++ backend/app/analysis/parts_of_speech.py
@@ -183,7 +183,6 @@ def get_word_definition(word_list, pos):
                               if word not in meaning]
         else:
             word_def[word] = []
-    print(word_def)
     return word_def
