+++ backend/app/views.py
@@ -21,6 +21,9 @@ from .analysis.anagrams import (
     get_anagrams,
     get_letter_freq,
 )
+from .analysis.textdata import (
+    get_text_data,
+)
 
 
 @api_view(['GET'])
@@ -116,6 +119,7 @@ def add_text(request):
     body = json.loads(request.body.decode('utf-8'))
     new_text_obj = Text(title=body['title'], content=body['content'])
     new_text_obj.save()
+    get_text_data(new_text_obj)
     serializer = TextSerializer(new_text_obj)
     return Response(serializer.data)
